move pu with space test, remove broken-pu related test - already covered by MultipleWriteMappingsForFieldTest (in core.test)
