+++ jpa/eclipselink.jpa.test/src/it/resources/eclipselink-advanced-properties/persistence.xml
@@ -46,12 +46,4 @@
              <property name="eclipselink.profiler" value="NoProfiler"/>
         </properties>
     </persistence-unit>
-    <!-- Bug 210280 testing for URL encoded char hex(20)-space and hex(C3A1)-multibyte handling in the PU name -->
-    <persistence-unit name="A JPAADVProperties pu with spaces in the name" transaction-type="RESOURCE_LOCAL">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <exclude-unlisted-classes>false</exclude-unlisted-classes>
-        <properties>
-             <property name="eclipselink.profiler" value="NoProfiler"/>
-        </properties>
-    </persistence-unit>
 </persistence>
